refactor: 회원 정보 조회 기능 수정

+++ backend/src/main/java/group/ict/sosservice/authentication/handler/CustomLogoutHandler.java
@@ -23,7 +23,7 @@ public class CustomLogoutHandler implements LogoutHandler {
     ) {
         final HttpSession session = request.getSession();
 
-        if (session != null) {
+        if (session != null && authentication != null) {
             tokenRepository.removeUserTokens(authentication.getName());
             session.invalidate();
         }
